Meilleur emplacement pour SERVICE_STATUS_PRESENTATION
