+++ src/lib/components/services/service-state-update-select.svelte
@@ -15,14 +15,61 @@
   } from "$lib/services";
 
   import { SERVICE_STATUSES, type Service } from "$lib/types";
-  import {
-    SERVICE_STATUS_PRESENTATION,
-    getAvailableOptionsForStatus,
-  } from "$lib/utils/service";
+  import { getAvailableOptionsForStatus } from "$lib/utils/service";
   import { validate } from "$lib/validation";
   import { serviceSchema } from "$lib/schemas/service.js";
-  import { arrowDownSIcon, arrowUpSIcon, deleteBinIcon } from "$lib/icons";
+  import {
+    arrowDownSIcon,
+    arrowUpSIcon,
+    deleteBinIcon,
+    earthFill,
+    fileEditFill,
+    folderFill,
+    draftFill,
+  } from "$lib/icons";
   import { clickOutside } from "../use/click-outside";
+
+  type ServiceStatusPresentation = {
+    bgClass: string;
+    iconClass: string;
+    hoverBgClass: string;
+    icon: string;
+    label: string;
+  };
+
+  export const SERVICE_STATUS_PRESENTATION: Record<
+    SERVICE_STATUSES,
+    ServiceStatusPresentation
+  > = {
+    [SERVICE_STATUSES.SUGGESTION]: {
+      bgClass: "bg-service-violet",
+      iconClass: "text-service-violet-darker",
+      hoverBgClass: "bg-service-violet-dark",
+      icon: fileEditFill,
+      label: "Suggestion",
+    },
+    [SERVICE_STATUSES.DRAFT]: {
+      bgClass: "bg-service-orange",
+      iconClass: "text-service-orange-darker",
+      hoverBgClass: "bg-service-orange-dark",
+      icon: draftFill,
+      label: "Brouillon",
+    },
+    [SERVICE_STATUSES.PUBLISHED]: {
+      bgClass: "bg-service-green",
+      iconClass: "text-service-green-darker",
+      hoverBgClass: "bg-service-green-dark",
+      icon: earthFill,
+      label: "Publié",
+    },
+    [SERVICE_STATUSES.ARCHIVED]: {
+      bgClass: "bg-service-gray",
+      iconClass: "text-gray-darker",
+      hoverBgClass: "bg-service-gray-dark",
+      icon: folderFill,
+      label: "Archivé",
+    },
+  };
 
   export let service: Service;
   export let servicesOptions;
